requirements.txt
certifi==2020.12.5
chardet==4.0.0
Django==1.11
idna==2.10
m3-builder==1.2.0
m3-core==2.2.24
m3-django-compat==1.9.2
m3-objectpack==2.2.25
m3-ui==2.2.96
pytz==2021.1
requests==2.25.1
six==1.15.0
urllib3==1.26.4
